+++ projects/common/src/lib/table/table.component.ts
@@ -56,6 +56,7 @@ export class PluggularTableComponent implements OnInit {
   @Input() showPagination = true;
   @Input() tableRowStyle = '';
   @Input() page = 0;
+  @Input() id = 'plg-table';
   @Output() hasSortOptionChanged = new EventEmitter<string>();
   @Output() hasPageChanged = new EventEmitter<number>();
   @Output() hasPageLimitChanged = new EventEmitter<string>();